requirements.txt
toml
json
pandas
streamlit
google-cloud-firestore
google-auth-oauthlib